Restructure translation directory
Restructure the translation directory to include translation
subdirectories for both the Macroni and SafeC dialects. We do this
we will be translating directly to the SafeC dialect now, and not
converting fromn the Macroni dialect.
