+++ include/macroni/Translation/CMakeLists.txt
@@ -1,1 +1,2 @@
-# Copyright (c) 2023-present, Trail of Bits, Inc.
+add_subdirectory(Macroni)
+add_subdirectory(Safety)
