+++ include/macroni/Translation/Macroni/MacroniMetaGenerator.hpp
