+++ lib/macroni/Translation/CMakeLists.txt
@@ -1,4 +1,2 @@
-add_library(macroni_translation_api STATIC MacroniMetaGenerator.cpp
-                                           MacroniVisitor.cpp)
-target_link_libraries(macroni_translation_api PRIVATE macroni_settings)
-add_library(macroni::translation_api ALIAS macroni_translation_api)
+add_subdirectory(Macroni)
+add_subdirectory(Safety)
