+++ lib/macroni/Translation/Macroni/CMakeLists.txt
@@ -1,0 +1,4 @@
+add_library(macroni_translation_api STATIC MacroniMetaGenerator.cpp
+                                           MacroniVisitor.cpp)
+target_link_libraries(macroni_translation_api PRIVATE macroni_settings)
+add_library(macroni::translation_api ALIAS macroni_translation_api)
